move _fdrx -> fdr._fdrx
