+++ ruff-lint-strict.toml
@@ -12,9 +12,9 @@ exclude = [
     "alphadia/search_plan.py",
     "alphadia/search_step.py",
     "alphadia/utils.py",
-    "alphadia/_fdrx/stats.py",
-    "alphadia/_fdrx/plotting.py",
-    "alphadia/_fdrx/base.py",
+    "alphadia/fdr/_fdrx/stats.py",
+    "alphadia/fdr/_fdrx/plotting.py",
+    "alphadia/fdr/_fdrx/base.py",
     "alphadia/constants/**.py",
     "alphadia/libtransform/**.py",
     "alphadia/outputtransform/**.py",
