+++ tests/unit_tests/fdr/_fdrx/test_plotting.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 import matplotlib.pyplot as plt
 import numpy as np
 
-from alphadia._fdrx.plotting import (
+from alphadia.fdr._fdrx.plotting import (
     _plot_fdr_curve,
     _plot_roc_curve,
     _plot_score_distribution,
